+++ src/api/post.api.ts
@@ -1,0 +1,22 @@
+import { instance } from "../lib/Axios";
+
+export const handelGetPost = async () => {
+  const url: string = "v1/posting/";
+  const { data }: any = await instance.get(url);
+
+  return data;
+};
+
+export const handelPostPosting = async (req: object) => {
+  const url: string = "v1/posting/";
+  const { data }: any = await instance.post(url, req);
+
+  return data;
+};
+
+export const handelDeletePost = async (req: object) => {
+  const url: string = "v1/posting/";
+  const { data }: any = await instance.delete(url, { data: req });
+
+  return data;
+};
