--- src/lib/api/index.api.ts
@@ -1,8 +1,0 @@
-import { request } from "../Axios";
-
-export const handleGetUrl = async () => {
-  const url: string = "v1/url/";
-  const { data }: any = await request.get(url);
-
-  return data;
-};
